backend_service/requirements.txt
django
djangorestframework
requests
firebase-admin
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-django
